requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --no-hashes --no-dev --no-group mcp --output-file requirements.txt
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
    #   watchfiles
argon2-cffi==25.1.0
    # via study-dashboard-api
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
attrs==26.1.0
    # via
    #   jsonschema
    #   referencing
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
cffi==2.0.0
    # via
    #   argon2-cffi-bindings
    #   cryptography
click==8.3.2
    # via uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   uvicorn
cryptography==46.0.7
    # via pyjwt
deprecation==2.1.0
    # via postgrest
fastapi==0.136.0
    # via study-dashboard-api
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   mcp
    #   postgrest
    #   study-dashboard-api
httpx-sse==0.4.3
    # via mcp
hyperframe==6.1.0
    # via h2
idna==3.11
    # via
    #   anyio
    #   httpx
    #   yarl
itsdangerous==2.2.0
    # via study-dashboard-api
jsonschema==4.26.0
    # via mcp
jsonschema-specifications==2025.9.1
    # via jsonschema
mcp==1.27.0
    # via study-dashboard-api
multidict==6.7.1
    # via yarl
packaging==26.1
    # via deprecation
pillow==12.2.0
    # via study-dashboard-api
postgrest==2.28.3
    # via study-dashboard-api
propcache==0.4.1
    # via yarl
pycparser==3.0 ; implementation_name != 'PyPy'
    # via cffi
pydantic==2.13.2
    # via
    #   fastapi
    #   mcp
    #   postgrest
    #   pydantic-settings
    #   study-dashboard-api
pydantic-core==2.46.2
    # via pydantic
pydantic-settings==2.13.1
    # via
    #   mcp
    #   study-dashboard-api
pyjwt==2.12.1
    # via mcp
pymupdf==1.27.2.2
    # via study-dashboard-api
python-dateutil==2.9.0.post0
    # via study-dashboard-api
python-dotenv==1.2.2
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.26
    # via
    #   mcp
    #   study-dashboard-api
pywin32==311 ; sys_platform == 'win32'
    # via mcp
pyyaml==6.0.3
    # via uvicorn
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
sse-starlette==3.3.4
    # via mcp
starlette==1.0.0
    # via
    #   fastapi
    #   mcp
    #   sse-starlette
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   mcp
    #   pydantic
    #   pydantic-core
    #   referencing
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   mcp
    #   pydantic
    #   pydantic-settings
tzdata==2026.1
    # via study-dashboard-api
uvicorn==0.44.0
    # via
    #   mcp
    #   study-dashboard-api
uvloop==0.22.1 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via uvicorn
yarl==1.23.0
    # via postgrest
